// types/resources.ts
export enum ResourceType {
  BLOOD = 'BLOOD',
  SOULS = 'SOULS',
  BONE = 'BONE',
  OBSIDIAN = 'OBSIDIAN',
  IRON = 'IRON',
  ESSENCE = 'ESSENCE',
  VOID_CRYSTALS = 'VOID_CRYSTALS',
  INFLUENCE = 'INFLUENCE',
  DARK_WEAPONS = 'DARK_WEAPONS',
  DARK_ARMOR = 'DARK_ARMOR',
  SLAVES = 'SLAVES',
  REFINED_OBSIDIAN = 'REFINED_OBSIDIAN',
  CONCENTRATED_ESSENCE = 'CONCENTRATED_ESSENCE',
  GOLD = 'GOLD'
}